Added source function to determine what color each box should be in a guessed word.
Added function correctLetters in check-words.mjs. correctLetters take two strings, guess and answer, and determines
which letters are in the correct spot, which are somewhere else in the word, and which aren't in the word.
The function returns an array of strings representing what color each box should be in the guessed word.